feat(storage): Implement `Compact` for `OpTxEnvelope` from `op_alloy`
