+++ crates/storage/codecs/src/alloy/transaction/optimism.rs
@@ -1,12 +1,22 @@
 //! Compact implementation for [`AlloyTxDeposit`]
 
-use alloy_consensus::constants::EIP7702_TX_TYPE_ID;
-use crate::Compact;
-use alloy_primitives::{Address, Bytes, TxKind, B256, U256};
-use op_alloy_consensus::{OpTxType, OpTypedTransaction, TxDeposit as AlloyTxDeposit};
+use crate::{
+    alloy::transaction::envelope::impl_compact_tx_envelope,
+    generate_tests,
+    txtype::{
+        COMPACT_EXTENDED_IDENTIFIER_FLAG, COMPACT_IDENTIFIER_EIP1559, COMPACT_IDENTIFIER_EIP2930,
+        COMPACT_IDENTIFIER_LEGACY,
+    },
+    Compact,
+};
+use alloy_consensus::{
+    constants::EIP7702_TX_TYPE_ID, Signed, TxEip1559, TxEip2930, TxEip7702, TxLegacy,
+};
+use alloy_primitives::{Address, Bytes, PrimitiveSignature, Sealed, TxKind, B256, U256};
+use bytes::BufMut;
+use lazy_static::lazy_static;
+use op_alloy_consensus::{OpTxEnvelope, OpTxType, OpTypedTransaction, TxDeposit as AlloyTxDeposit};
 use reth_codecs_derive::add_arbitrary_tests;
-use crate::txtype::{COMPACT_EXTENDED_IDENTIFIER_FLAG, COMPACT_IDENTIFIER_EIP1559, COMPACT_IDENTIFIER_EIP2930, COMPACT_IDENTIFIER_LEGACY};
-use crate::generate_tests;
 
 /// Deposit transactions, also known as deposits are initiated on L1, and executed on L2.
 ///
@@ -68,7 +78,6 @@ impl Compact for AlloyTxDeposit {
         (alloy_tx, buf)
     }
 }
-
 
 impl crate::Compact for OpTxType {
     fn to_compact<B>(&self, buf: &mut B) -> usize
@@ -159,5 +168,78 @@ impl Compact for OpTypedTransaction {
         }
     }
 }
+
+impl crate::alloy::transaction::envelope::ToTxCompact for OpTxEnvelope {
+    fn to_tx_compact(&self, buf: &mut (impl BufMut + AsMut<[u8]>)) {
+        match self {
+            Self::Legacy(tx) => tx.tx().to_compact(buf),
+            Self::Eip2930(tx) => tx.tx().to_compact(buf),
+            Self::Eip1559(tx) => tx.tx().to_compact(buf),
+            Self::Eip7702(tx) => tx.tx().to_compact(buf),
+            Self::Deposit(tx) => tx.to_compact(buf),
+        };
+    }
+}
+
+impl crate::alloy::transaction::envelope::FromTxCompact for OpTxEnvelope {
+    type TxType = OpTxType;
+
+    fn from_tx_compact(
+        buf: &[u8],
+        tx_type: OpTxType,
+        signature: PrimitiveSignature,
+    ) -> (Self, &[u8]) {
+        match tx_type {
+            OpTxType::Legacy => {
+                let (tx, buf) = TxLegacy::from_compact(buf, buf.len());
+                let tx = Signed::new_unhashed(tx, signature);
+                (Self::Legacy(tx), buf)
+            }
+            OpTxType::Eip2930 => {
+                let (tx, buf) = TxEip2930::from_compact(buf, buf.len());
+                let tx = Signed::new_unhashed(tx, signature);
+                (Self::Eip2930(tx), buf)
+            }
+            OpTxType::Eip1559 => {
+                let (tx, buf) = TxEip1559::from_compact(buf, buf.len());
+                let tx = Signed::new_unhashed(tx, signature);
+                (Self::Eip1559(tx), buf)
+            }
+            OpTxType::Eip7702 => {
+                let (tx, buf) = TxEip7702::from_compact(buf, buf.len());
+                let tx = Signed::new_unhashed(tx, signature);
+                (Self::Eip7702(tx), buf)
+            }
+            OpTxType::Deposit => {
+                let (tx, buf) = op_alloy_consensus::TxDeposit::from_compact(buf, buf.len());
+                let tx = Sealed::new(tx);
+                (Self::Deposit(tx), buf)
+            }
+        }
+    }
+}
+
+lazy_static! {
+    static ref DEPOSIT_SIGNATURE: PrimitiveSignature =
+        PrimitiveSignature::new(U256::ZERO, U256::ZERO, false);
+}
+
+trait Envelope {
+    fn signature(&self) -> &PrimitiveSignature;
+}
+
+impl Envelope for OpTxEnvelope {
+    fn signature(&self) -> &PrimitiveSignature {
+        match self {
+            Self::Legacy(tx) => tx.signature(),
+            Self::Eip2930(tx) => tx.signature(),
+            Self::Eip1559(tx) => tx.signature(),
+            Self::Eip7702(tx) => tx.signature(),
+            Self::Deposit(_) => &DEPOSIT_SIGNATURE,
+        }
+    }
+}
+
+impl_compact_tx_envelope!(OpTxEnvelope);
 
 generate_tests!(#[crate, compact] OpTypedTransaction, OpTypedTransactionTests);
